Inclusion of a method for selecting the best model and number of parameters for use in the plots
Assumes featRank_D is the best combination of importance-contribution weighting at 0.5.  Finds the performance values with the maximum score and pulls the number of features.  The number of features for training the subsequent models as well as the figures reflects this number of features.  Assumes that only one set of parameters will have the maximum score.  The final model is still assumed to be based on featRank_D.