Uniform variables for different weighting of import-contrib
Created variables to set for the number of features, feature list, and column names for the best features from a given feature ranking weighting of importance and contribution.  This enables quick copy-paste of figure generating blocks editing only figure title descriptions and file naming.  This update was applied to 0.75 and 1 weights, not 0.5 yet.  File naming updated to include a "_w#" where the number is the importance-contribution weighting.