Improved color mapping for attribution plots
Adjusted the ranges for the color mapping for the performance and attribution plots for the different weighted systems to better show the data.  Integrated gradients for one model all use the same color map scale.  Consider changing the color map scale to be homogeneous across all classifiers for easier comparison between plots.