Creation of iterative feature reduction
After 0.3 feature reduction, collect the top 1/3 best features from each of the 0.5, 0.75, and 1 weighting of importance-contribution.  Take the unique list of features and create a new X_train_all dataframe.  Run the feature selection and data visualization using 0.1 feature reduction.

Also improvement in a few blocks of figure generating code of using the more universal variables for easier modification.

Have not tested to see if it runs.