Column selection corrected
Updated column selection for the second iteration of the feature reduction to use the indices from the original list.  X_train_all is a slice of the already scaled data from the first pass.  This prevents needing to select the columns, and reapply the StandardScaler.

File names were also corrected as the length was to long for the directory.

Also corrected the second round of hyper parameter tuning as some cells were set to markdown instead of code preventing appropriate variable updates.