Correcting data filter and running code
In the "collecting the lab columns we need" section of the code, there are two lines of code that correct the data from Anisocytosis Result and Macrocytes Result to remove the + after the numeric score.  Comment out these lines since those labs are not included in the top 60 labs.

Also some minor directory updates.