Cmap by amyloid status for PatientDataCompleteness
Upated patient data completeness visualizations to map the amyloid status of a patient to a different colormap to use in the same figure.  Box plots added for the average instances of a lab across amyloid positive and negative patients.

This can be used for baseline code in the FeatureBasedListComparison describing the features selected by a model.