Creating script for cv_set_for_rank test
CV_set_for_rank is a component of the greedy_backward_selection process that determines if the contribution scores are determined by a set of data from training or test.  The default is test data, we are setting it to training data to see if this influences the performance of the models as this could be a source of bias.