Run with CV_set_for_rank set to train
Run of the scrip where CV_set_for_rank for greedy_backward_selection is set to train instead of test.  The script was also updated such that the colormapping center is set to 0.  This improves the figure interpretations.  Another set of figures was included for a smaller slice of features in the performance plots AUC and F1 score.