Updating the swimmer plots for the poster
Creation of swimmer plots where one figure includes both the amyloid positive and negative patients. The diagnosis (cancer) has been removed from the plot.  Zero values are not plotted in the swimmer.
There is options for plotting the histogram for the lab value adjacent to the swimmer plot.

Limitations:
There is some crash in the titles that will need to be manually adjusted for use in posters