Multiple Lab swimmer broken
Attempted to create dictionaries so that each  lab would have a different marker and color.
The density of the data makes this very unclear on the plots.
A combination marker would make this a little easier, but scanning for both labs for a combination marker is challenging in the code at this time.