Feature correlations and new dataframe
Created a notebook for examining the correlational relationship between features derived from fits of the data to model equations.

Also updated a notebook for creation of new dataframe with encoding for categorical variables and retention of nan for missing values. MD notes descriptive data retains penalty score of -1 to handle the missing data from T-01 but will be updated soon.  There is also a new set of functions to preprocess data for fits to model equations.