Analysis of data length by lab
Patients with data length less than 2 were dropped from the patient list.

A value_counts() is performed on the dataframe describing the number of time points per lab per patient.  The columns describing the data lengths are then sorted.  The data types are then sorted by the frequency of data length 0 and 1.