Running mistic1 on CompLabswoEcho 241017
Running an older version of mistic using the new dataframe CompLabswoEcho from 2024-10-17.  This set of labs drops patients with data length 1 and excludes labs with data length 0.  BNP and troponin were not included as a result.  Descriptive statistics of data length 1 were set to the original value.  The model performs well, better than the "equivalent" model run with the old penalty scoring system.