Inclusion of cardiac markers
Dropped Pts G01 and J01 as they lack cardiac markers. Dropped K-01 for low amounts of data.  Dropped L-02 for lacking kappa light chain measures.  Adjusted the scoring for clinical impression of disease to include nan for missing or incomplete data. This resolves itself as the features are the descriptive stats of this data. Set the residual error for fits of data length 1 to the value itself.

Constructed the mistic version 1 to go with this to test the performance of this data before moving to mistic 2