Investigation of md notes for fit
The ECOG and Karnofsky values can be fit to model equations.  Clinical impression of disease is to short to fit to model of equations and is often highly missing.  Patient T-01 lacks physician note information, so -1 was filled in for all note data features of this patient.  In order to get descriptive stats for the note data fits, data type must be forced to float.

Constructed an older version of mistic to test this data to see if it is helpful in the interpretability of the classifier.  If successful, the rest of the model systems will be built and a new mistic classifier will be constructed.