Run of mistic 1 with mdFits and red005
Ran the old classifier with the dataframe from 2024-10-18 using fits of ECOG and Karnofsky scores.  Reduced the reduction factor 0.1 to 0.05. The performance is good, but there are so many features and many of them redundant. This feature list might get shorter by further reducing the reduction factors.  But, first exploring different data groups.