Created mistic 2 with desc only and params only features
Created versions of mistic 2 for training SVMs on the feature sets containing either descriptive stats only or parameters of fit only.  The separate feature sets and separate parameter sets are set to False to best approximate old mistic.  A single model feature list and set of parameters will result for each weighting of importance-contribution.  The parameters and feature lists are exported to an excel file with different sheets for features and performance per weighting.  The performance figures export to file.  The attribution figures are still not exporting correctly.  These figures are best handled from the pickled svm_set_ranks object on the local machine.  Feature reduction set to 0.05.  5-fold cross validation is used.  There is still room for improvement by holding out the imbalanced patient to use as a test of the model.