Attempt to load from pickle
Attempting to load svm set from pickle to work on figures.  However, files names were to long, the mistic package wont import fully, and the pickles "ran out of input".  May need to rerun the SVMs with shorter file names to see if data was lost in needing to rename the files