Testing pickle and max_iter limit
The descOnly model was failing at the weight of 0.5 and feature number 3.  Set a limit on the maximum number of iterations to 100000.  Also testing the pickling and unpickling of the svm_set_ranks.  This was failing as a fresh run of the code when I was working on creating figures from executed files.