Figure file extension options and histograms
Updated the 2D histograms for the linear kernels.  There is an error when running statistics that might be related to package differences between my home machine and work machine.

Updates to the patient functions file to enable specification of the file extension for figures.  The default is set to tif.