Alt models correct data split and LGBM solved
Alternative models were not using the correct data split or reporting of the F1 score.  F1 score updated to be binary and represent our binary classification.  The LGBM model was updated to have smaller minimum data sizes for a leaf to enable feature reduction.  The data trimming to the top features from the SVM model to retest the alternative models was updated to match the data selection done for the pairplots (but on the transformed data).  This reduces the need to re-import the data or accidentally change the data from the original processing.  Alternative model performance is not output to an excel file.