Updated stats notebook
Calculates T stats for EchoAdj feature list (removes the penalty scores first), identifies significant differences, and plots the box plots for the data.  Also calculates T stats for DescFits data based on CA status.  Next is to do the box plots based on CA status and check for spearman's correlation with the data sets.  Still struggling to run these stats in GraphPad since there is so much data transformation that occurs and hundreds of features.