EchoAdj and descriptive stats
Collected slice of patient data for imaging adjacent labs by calculating difference between imaging date and time of labs in days.  Took the absolute value and selected the minimum value from the list.  When a patient doesn't have an echo, the cardiac MRI date is used.  Some of the patients with unknown amyloid status lack imaging data entirely, so only the patients with known amyloid status were indexed.

0 was substituted for values below the measurable threshold for troponin and BNP.  Descriptive statistics were then calculated for the list of filtered labs across patients with known amyloid status.  The statistics dataframes were flattened using stack and the multi-indexing resolved using the _ delimiter.  Data of length 1 had std and other non-counts statistics set to the one recorded value (listed as "mean" in the describe function for data of length 1)

Next set of work is to gather descriptive statistics for the notes data and to encode the treatment information.