Completed making sheets and figures
Completed compiling sheets, handling dummy variables, plotting example fits, and plotting event plot of the data.  Also created a new SVM file for the descOnly datasheet to test on HPG.  The SVM file has an addition for plotting the correlation matrices of the selected features.  There is still room to improve that script and have the features appear in ranked order.