LinearOnly and ExpOnly
Created versions of the DescFits and ParamsOnly that contain only linear fits or only exponential fits.  These models will help with determining if the models selecting both linear and exponential of the same fit feature is informative or not to the model.

Also added a way to plot all 44 patients data and fits for each lab in grids.  Exporting these images to tifs for analysis later.  Exporting the event plot to file as well.