Run of lin/exp only models
Runs of models using only linear or exponential fits of the data.

Added a section for random forest model on its top features from the all features model where the number of features corresponds to the number of features selected by the SVM.  This is hard coded based on how we already know the SVM behaves in feature selection.