Iterative feature selection for DescFits linear
Testing a second round of feature selection from the 17 features selected by the DescFits model with linear kernel and 0.25 feature reduction.  Still testing a full range of weightings for importance and contribution.  Looking to see if the second round of feature selection reduces the selection of highly correlated features in the set.