Runs 17 features and 1 intercept
Testing iterative feature reduction from the initial best features list of 17 from descFits linear kernel feature reduction 0.25.

Testing using only the linear intercept for the fits.  Went through several feature reduction rates.  Still concerned for over-fitting as the number of features is to close to the number of patients in the dataset.