Data at time
To get the lab value and the relative time for different time based conditions: time closest to cardiac imaging, time closest to date of oncologic diagnosis, and first recorded value. Save this information to a sheet

To check if the list of labs changes if we filter for only times greater than 0.

The list of labs does not change when we limit time to greater than or equal to 0.  The completeness of patient data however may change.  It appears that patients with a low number of labs are unlikely to lose their data necessary to form a fit.