Second round of feature selection
Added block of code to pull list of features from a spreadsheet containing all of the different models, then perform feature selection and training on that list of features.  Has the ability to pull from a sheet for the features selected by this model before as well as older versions of these models under old features.

Improved the directory to put the model outputs in a sub folder on hpg.