By feature reduction and counts
Looking at features for one model across rank weights for one feature reduction rate at a time.  Added in section of code to look at the number of times a feature or underlying original data was used in the grouping of models (many rank weights, one feature reduction rate).  Next is to improve the data visualization and export.