random seed 4200, feature reduction 0.25
Testing models using random seed 4200, placing patient T-01 into the test set instead of the train set.  feature reduction 0.25.  The goal was to determine if moving the patient to the test set would enable keeping the note data as features, since note data is not selected as a feature by the models.