Clinical Split SVMS feature reduction 0.25
Time is restricted to 60 days prior to oncologic diagnosis.  Physician note data is not used as a feature in the models but is used to form a clinically relevant test-train split.
The 1/5 test-train split is done using the distribution of patient mean KPS for functional status.

Feature reduction set to 0.25.