pairplots in stats visuals
hard coded a section in stats visuals for the pair plot with the rank ordered features from DescFits red 025 feature list.

Updated the file directory and file names to reflect the clinical split of the patients.

Next is to update the stats outputs to reflect the current datasets for plotting different aspects of the data.