Feature summary across models
Figures for summarizing the feature selection across SVM models including the original data selected and transformations used.  There are also sections of code for correlation tables of selected features.

Could add a section for correlation of all features selected across all models to see what patterns the models are finding.

There are also some updates to the figure outputs for the generation of the dataframes: plots of the labs across time for the test patients, numbers in heat maps for easier analysis.