IG plots and SNAFU
Runs of the new integrated gradient plots and a correction of the feature order list to match the full list of features (it was being altered by the number of features plotted in the abridged SHAP plot).

Then I found a change in the performance and selected models.  This is a change from the models run on 7/14.  Unfortunately, the exported outputs are overwritten at this time.  Thought is that there is a change in HPG hardware that is causing this change.