IG Plot markers and IG table for test
Improved the markers for the IG plot to pull from a unique list of the categorical variable of shape.  Created a loop to generate IG plots from a list of paired indices.  Updated the IG table clustermap to work form a subset of the original integrated gradients calculations so that we don't have to recompute it.  The EchoAdj model still needs updated and models need rerun for stability and versioning.