Plot of support vector counts
Plot of support vector selection across ensembles as well as their decision value and true CA status.  The tables were exported for use in comparing the between models.  An additional notebook will be made to create that graphic.

File dirrectories shorted and updated to 0728 for ease of location in the drive.